mysql ２回目

+++ db/migrate/20250313085103_add_user_id_to_visit_records.rb
@@ -1,6 +1,6 @@
 class AddUserIdToVisitRecords < ActiveRecord::Migration[7.1]
   def change
-    add_column :visit_records, :user_id, :integer, null: true
+    add_column :visit_records, :user_id, :bigint, null: true
     add_index :visit_records, :user_id
     add_foreign_key :visit_records, :users, on_delete: :cascade
   end
